Fix OTP-26 compilation
