+++ test/typed_struct_test.exs
@@ -1,6 +1,7 @@
 defmodule TypedStructTest do
   use ExUnit.Case
 
+  import TypedStructs
   import ExUnit.CaptureIO
 
   ############################################################################
@@ -28,9 +29,6 @@ defmodule TypedStructTest do
       field :field, term()
     end
   end
-
-  # Standard struct name used when comparing generated types.
-  @standard_struct_name TypedStructTest.TestStruct
 
   ############################################################################
   ##                             Standard cases                             ##
@@ -170,50 +168,4 @@ defmodule TypedStructTest do
     assert type1 == type2
   end
 
-  ############################################################################
-  ##                                Helpers                                 ##
-  ############################################################################
-
-  # Extracts the first type from a module.
-  defp extract_first_type(bytecode, type_keyword \\ :type)
-
-  defp extract_first_type(bytecode, type_keyword) when is_binary(bytecode) do
-    case Code.Typespec.fetch_types(bytecode) do
-      {:ok, types} -> Keyword.get(types, type_keyword)
-      _ -> nil
-    end
-  end
-
-  defp extract_first_type(module, type_keyword) when is_atom(module) do
-    {_, bytecode, _} = :code.get_object_code(module)
-    extract_first_type(bytecode, type_keyword)
-  end
-
-  defp standardize_first_type(module, type_keyword \\ :type)
-       when is_atom(module) do
-    extract_first_type(module, type_keyword)
-    |> standardise(module)
-  end
-
-  # Standardises a type (removes line numbers and renames the struct to the
-  # standard struct name).
-  defp standardise(type_info, struct)
-
-  defp standardise({name, type, params}, struct) when is_tuple(type),
-    do: {name, standardise(type, struct), params}
-
-  defp standardise({:type, _, type, params}, struct),
-    do: {:type, :line, type, standardise(params, struct)}
-
-  defp standardise({:remote_type, _, params}, struct),
-    do: {:remote_type, :line, standardise(params, struct)}
-
-  defp standardise({:atom, _, struct}, struct),
-    do: {:atom, :line, @standard_struct_name}
-
-  defp standardise({type, _, litteral}, _struct),
-    do: {type, :line, litteral}
-
-  defp standardise(list, struct) when is_list(list),
-    do: Enum.map(list, &standardise(&1, struct))
 end
